requirements.txt
# requirements.txt
face_recognition
numpy
opencv-python
requests
Pillow
boto3
